Combined external and internal supplier invoices to use the same page
