+++ src/pages/index.js
@@ -11,7 +11,6 @@ import { CustomerInvoicePage,
          checkForFinaliseError as checkForCustomerInvoiceFinaliseError,
        } from './CustomerInvoicePage';
 import { CustomerInvoicesPage } from './CustomerInvoicesPage';
-export { FirstUsePage } from './FirstUsePage';
 import { MenuPage } from './MenuPage';
 import { PageContainer } from './PageContainer';
 import { RequisitionsPage } from './RequisitionsPage';
@@ -25,11 +24,11 @@ import { StocktakeEditPage,
 import { StocktakeManagePage } from './StocktakeManagePage';
 import { StocktakesPage } from './StocktakesPage';
 import { SupplierInvoicePage } from './SupplierInvoicePage';
-import { SupplierInvoicesPage } from './SupplierInvoicesPage';
+import { SupplierInvoicesPage,
+         checkForFinaliseError as checkForSupplierInvoiceError,
+       } from './SupplierInvoicesPage';
 import { RealmExplorer } from './RealmExplorer';
-import { ExternalSupplierInvoicePage,
-         checkForFinaliseError as checkForExternalSupplierInvoiceError,
-       } from './ExternalSupplierInvoicePage';
+export { FirstUsePage } from './FirstUsePage';
 
 export const PAGES = {
   customer: (props) => <PageContainer page={CustomerPage} {...props} />,
@@ -47,7 +46,6 @@ export const PAGES = {
   stocktakes: (props) => <PageContainer page={StocktakesPage} {...props} />,
   supplierInvoice: (props) => <PageContainer page={SupplierInvoicePage} {...props} />,
   supplierInvoices: (props) => <PageContainer page={SupplierInvoicesPage} {...props} />,
-  externalSupplierInvoice: (props) => <PageContainer page={ExternalSupplierInvoicePage} {...props} />,
 };
 
 export const FINALISABLE_PAGES = {
@@ -57,7 +55,7 @@ export const FINALISABLE_PAGES = {
     finaliseText: 'finalise_supplier_invoice',
   },
   externalSupplierInvoice: {
-    checkForError: checkForExternalSupplierInvoiceError,
+    checkForError: checkForSupplierInvoiceError,
     recordType: 'Transaction',
     recordToFinaliseKey: 'transaction',
     finaliseText: 'finalise_supplier_invoice',
